Fixed exception reporting bug where certain exceptions were incorrectly reported as uncaught.
Improved the memory allocation strategy used during compilation to make running out of memory when compiling huge scripts less likely.

Optimized String.replace by avoiding the construction of certain sub strings.

Fixed bug in code generation for large switch statements on ARM.

Fixed bug that caused V8 to change the global object template passed in by the user.

Changed the API for creating object groups used during garbage collection.  Entire object groups are now passed to V8 instead of individual members of the groups.


git-svn-id: http://v8.googlecode.com/svn/trunk@968 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/utils.h
@@ -540,7 +540,8 @@ static inline void Store16(byte* ptr, uint16_t value) {
 #ifdef CAN_READ_UNALIGNED
   *reinterpret_cast<uint16_t*>(ptr) = value;
 #else
-  ptr[1] = value;
+  // Cast to avoid warning C4244 when compiling with Microsoft Visual C++.
+  ptr[1] = static_cast<byte>(value);
   ptr[0] = value >> 8;
 #endif
 }
